requirements.txt
streamlit
tensorflow
numpy
matplotlib
scipy
ezdxf
numpy-stl
fpdf
plotly
qrcode
pillow